backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
# AI 相關
langchain
langchain-qdrant
langchain-google-genai
langchain-community
qdrant-client
llama-cloud-services
ddgs
# File
pdfrw
python-docx